git-svn-id: https://svn.abls-habitat.fr/repo/Watchdog@4420 9f22af27-4b39-4cfe-9f85-4bed9fef4693

+++ prod/Watchdogd/Include/Json.h
@@ -34,7 +34,7 @@
  extern JsonBuilder *Json_create ( void );
  extern void Json_add_string ( JsonBuilder *builder, gchar *name, gchar *chaine );
  extern void Json_add_int ( JsonBuilder *builder, gchar *name, gint valeur );
- extern void Json_add_double ( JsonBuilder *builder, gchar *name, gfloat valeur );
+ extern void Json_add_double ( JsonBuilder *builder, gchar *name, gdouble valeur );
  extern void Json_add_bool ( JsonBuilder *builder, gchar *name, gboolean bool );
  extern gchar *Json_get_buf ( JsonBuilder *builder, gsize *taille_buf_p );
  extern JsonNode *Json_get_from_string ( gchar *chaine );
